image-report-system/requirements.txt
numpy>=1.26.0,<2.0
opencv-python>=4.9.0
pillow>=10.2.0
torch>=2.2.0
segmentation-models-pytorch>=0.3.3
onnxruntime>=1.17.0
python-docx>=1.1.0
scikit-learn>=1.4.0
pandas>=2.2.0
pyyaml>=6.0.1
pytest>=8.0.0
